render utils for api and server tests
