+++ __tests__/api/title.test.js
@@ -1,20 +1,8 @@
-import React, { StrictMode } from 'react';
-import ReactDOM from 'react-dom';
+import React from 'react';
 import Helmet from '../../src';
-import Provider from '../../src/Provider';
+import { render } from './utils';
 
 Helmet.defaultProps.defer = false;
-
-const mount = document.getElementById('mount');
-
-const render = node => {
-  ReactDOM.render(
-    <StrictMode>
-      <Provider>{node}</Provider>
-    </StrictMode>,
-    mount
-  );
-};
 
 describe('title', () => {
   describe('API', () => {
